+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [1.2.3] - 2021-01-02
+
 ### Changed
 
 - [@lukyth], [@kddeisz] - Ensure if a ternary breaks into an `if..else..end` within a `command_call` node that parentheses are added.
@@ -1007,7 +1009,8 @@ would previously result in `array[]`, but now prints properly.
 
 - Initial release 🎉
 
-[unreleased]: https://github.com/prettier/plugin-ruby/compare/v1.2.2...HEAD
+[unreleased]: https://github.com/prettier/plugin-ruby/compare/v1.2.3...HEAD
+[1.2.3]: https://github.com/prettier/plugin-ruby/compare/v1.2.2...v1.2.3
 [1.2.2]: https://github.com/prettier/plugin-ruby/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/prettier/plugin-ruby/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/prettier/plugin-ruby/compare/v1.1.0...v1.2.0